api-dashboard/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Template engine
jinja2==3.1.2

# Static files
aiofiles==23.2.1

# Database drivers
asyncpg==0.29.0          # PostgreSQL async driver
motor>=3.4.0              # MongoDB async driver
pymongo>=4.6.0,<5.0.0     # MongoDB driver (compatible with motor 3.4+)
redis==5.0.1              # Redis client

# Kafka client
aiokafka==0.9.0
kafka==1.3.5  # For sync producer (DLQ)

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Data processing
pandas==2.1.4
numpy==1.25.2

# Monitoring and metrics
prometheus-client==0.19.0

# Logging
structlog==23.2.0

# Configuration
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# CORS support
fastapi-cors==0.0.6

# WebSocket support (included in FastAPI)
# Additional WebSocket utilities
websockets==12.0

# Security
cryptography>=41.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation
email-validator==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Type hints
mypy==1.7.1
types-redis==4.6.0.11

# Environment and deployment
gunicorn==21.2.0

# Utilities
click==8.1.7
rich==13.7.0